feat(api): Deprecate v1.0, move WS endpoints to separate file
