+++ app.py
@@ -7,6 +7,7 @@ from src.api.v1_0.action_creator.routes import action_creator_router_v1_0
 from src.api.v1_0.auth.routes import auth_router_v1_0
 from src.api.v1_0.health.routes import health_router_v1_0
 from src.api.v1_1.action_creator.routes import action_creator_router_v1_1
+from src.api.v1_1.action_creator.ws import action_creator_ws_router_v1_1
 from src.api.v1_2.action_creator.routes import action_creator_router_v1_2
 from src.api.v1_2.visualization.routes import visualization_router_v1_2
 from src.core.settings import current_settings
@@ -16,14 +17,14 @@ settings = current_settings()
 
 def register_api_v1_0(app: FastAPI) -> FastAPI:
     sub_app = FastAPI(
-        title="EOPro Action Creator API",
+        title="EOPro Action Creator API [DEPRECATED]",
         version="1.0.0",
         description="API for Action Creator.",
         debug=settings.environment.lower() in {"local", "dev"},
     )
-    sub_app.include_router(health_router_v1_0)
-    sub_app.include_router(auth_router_v1_0)
-    sub_app.include_router(action_creator_router_v1_0)
+    sub_app.include_router(health_router_v1_0, deprecated=True)
+    sub_app.include_router(auth_router_v1_0, deprecated=True)
+    sub_app.include_router(action_creator_router_v1_0, deprecated=True)
     app.mount("/api/v1.0", sub_app)
     return sub_app
 
@@ -38,6 +39,7 @@ def register_api_v1_1(app: FastAPI) -> FastAPI:
     sub_app.include_router(health_router_v1_0)
     sub_app.include_router(auth_router_v1_0)
     sub_app.include_router(action_creator_router_v1_1)
+    sub_app.include_router(action_creator_ws_router_v1_1, deprecated=True)
     app.mount("/api/v1.1", sub_app)
     return sub_app
